refactor: move feature type into a separate file
