+++ pkg/oomstore/types/types.go
@@ -22,26 +22,6 @@ type Entity struct {
 	ModifyTime  time.Time `db:"modify_time"`
 }
 
-type Feature struct {
-	ID          int16  `db:"id"`
-	Name        string `db:"name"`
-	GroupName   string `db:"group_name"`
-	ValueType   string `db:"value_type"`
-	DBValueType string `db:"db_value_type"`
-
-	Description string    `db:"description"`
-	CreateTime  time.Time `db:"create_time"`
-	ModifyTime  time.Time `db:"modify_time"`
-}
-
-type RichFeature struct {
-	Feature
-	EntityName string  `db:"entity_name"`
-	Category   string  `db:"category"`
-	Revision   *int64  `db:"revision"`
-	DataTable  *string `db:"data_table"`
-}
-
 type FeatureGroup struct {
 	ID               int16   `db:"id"`
 	Name             string  `db:"name"`
@@ -67,72 +47,6 @@ type Revision struct {
 	ModifyTime  time.Time `db:"modify_time"`
 }
 
-type FeatureList []*Feature
-type RichFeatureList []*RichFeature
-
-func (l *FeatureList) Len() int     { return len(*l) }
-func (l *RichFeatureList) Len() int { return len(*l) }
-
-func (l *FeatureList) Names() (names []string) {
-	for _, f := range *l {
-		names = append(names, f.Name)
-	}
-	return
-}
-
-func (l *FeatureList) Ids() (ids []int16) {
-	for _, f := range *l {
-		ids = append(ids, f.ID)
-	}
-	return
-}
-
-func (l *FeatureList) Filter(filter func(f *Feature) bool) (rs FeatureList) {
-	for _, f := range *l {
-		if filter(f) {
-			rs = append(rs, f)
-		}
-	}
-	return
-}
-
-func (l *RichFeatureList) Names() (names []string) {
-	for _, f := range *l {
-		names = append(names, f.Name)
-	}
-	return
-}
-
-func (l *RichFeatureList) Ids() (ids []int16) {
-	for _, f := range *l {
-		ids = append(ids, f.ID)
-	}
-	return
-}
-
-func (l *RichFeatureList) ToFeatureList() (rs FeatureList) {
-	for _, f := range *l {
-		rs = append(rs, f.AsFeature())
-	}
-	return
-}
-
-func (rf *RichFeature) AsFeature() *Feature {
-	if rf == nil {
-		return nil
-	}
-	return &rf.Feature
-}
-
-func (l *RichFeatureList) Filter(filter func(f *RichFeature) bool) (rs RichFeatureList) {
-	for _, f := range *l {
-		if filter(f) {
-			rs = append(rs, f)
-		}
-	}
-	return
-}
-
 type FeatureKV struct {
 	FeatureName  string
 	FeatureValue interface{}
@@ -156,46 +70,6 @@ func NewFeatureDataSet() FeatureDataSet {
 type EntityRowWithFeatures struct {
 	EntityRow
 	FeatureValues []FeatureKV
-}
-
-func (rf *RichFeature) String() string {
-	var revision, dataTable string
-
-	if rf.Revision == nil {
-		revision = "<NULL>"
-	} else {
-		revision = fmt.Sprint(*rf.Revision)
-	}
-
-	if rf.DataTable == nil {
-		dataTable = "<NULL>"
-	} else {
-		dataTable = *rf.DataTable
-	}
-
-	return strings.Join([]string{
-		fmt.Sprintf("Name:          %s", rf.Name),
-		fmt.Sprintf("Group:         %s", rf.GroupName),
-		fmt.Sprintf("Entity:        %s", rf.EntityName),
-		fmt.Sprintf("Category:      %s", rf.Category),
-		fmt.Sprintf("DBValueType:   %s", rf.DBValueType),
-		fmt.Sprintf("ValueType:     %s", rf.ValueType),
-		fmt.Sprintf("Description:   %s", rf.Description),
-		fmt.Sprintf("Revision:      %s", revision),
-		fmt.Sprintf("DataTable:     %s", dataTable),
-		fmt.Sprintf("CreateTime:    %s", rf.CreateTime.Format(time.RFC3339)),
-		fmt.Sprintf("ModifyTime:    %s", rf.ModifyTime.Format(time.RFC3339)),
-	}, "\n")
-}
-
-func (e *Entity) String() string {
-	return strings.Join([]string{
-		fmt.Sprintf("Name:          %s", e.Name),
-		fmt.Sprintf("Length:        %d", e.Length),
-		fmt.Sprintf("Description:   %s", e.Description),
-		fmt.Sprintf("CreateTime:    %s", e.CreateTime.Format(time.RFC3339)),
-		fmt.Sprintf("ModifyTime:    %s", e.ModifyTime.Format(time.RFC3339)),
-	}, "\n")
 }
 
 func (fg *FeatureGroup) String() string {
@@ -252,4 +126,14 @@ type RawFeatureValueRecord struct {
 type EntityRow struct {
 	EntityKey string `db:"entity_key"`
 	UnixTime  int64  `db:"unix_time"`
+}
+
+func (e *Entity) String() string {
+	return strings.Join([]string{
+		fmt.Sprintf("Name:          %s", e.Name),
+		fmt.Sprintf("Length:        %d", e.Length),
+		fmt.Sprintf("Description:   %s", e.Description),
+		fmt.Sprintf("CreateTime:    %s", e.CreateTime.Format(time.RFC3339)),
+		fmt.Sprintf("ModifyTime:    %s", e.ModifyTime.Format(time.RFC3339)),
+	}, "\n")
 }
